requirements.txt
accelerate==1.13.0
altair==6.0.0
annotated-doc==0.0.4
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
brotli==1.2.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
feedparser==6.0.12
filelock==3.25.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lxml==5.4.0
lxml_html_clean==0.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.17.0
networkx==3.6.1
newspaper4k==0.9.5
nltk==3.9.3
numpy==2.4.3
packaging==26.0
pandas==2.3.3
pillow==12.1.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.1
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-file==3.0.1
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
setuptools==82.0.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
textblob==0.19.0
tldextract==5.3.1
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
w3lib==2.4.0
watchdog==6.0.0
